trading_agent/learning/retraining.py
"""
Learning and model retraining.
Stubs for continuous model improvement.
"""

from abc import ABC, abstractmethod
from typing import List, Dict, Optional
from datetime import datetime
import pickle

from utils.types import Trade


class ModelTrainer(ABC):
    """Abstract base class for model trainers."""
    
    @abstractmethod
    def train(self, training_data: Dict) -> object:
        """Train a model."""
        pass
    
    @abstractmethod
    def save_model(self, model: object, path: str):
        """Save trained model."""
        pass
    
    @abstractmethod
    def load_model(self, path: str) -> object:
        """Load a trained model."""
        pass


class WalkForwardValidator:
    """
    Walk-forward validation for model development.
    Tests model on out-of-sample periods.
    """
    
    def __init__(
        self,
        training_window_days: int = 252,
        test_window_days: int = 30,
        step_days: int = 10,
    ):
        """
        Initialize walk-forward validator.
        
        Args:
            training_window_days: Size of training window
            test_window_days: Size of test window
            step_days: Step forward each iteration
        """
        self.training_window_days = training_window_days
        self.test_window_days = test_window_days
        self.step_days = step_days
    
    def generate_windows(self, start_date: datetime, end_date: datetime):
        """
        Generate walk-forward windows.
        
        Yields:
            Tuples of (training_start, training_end, test_start, test_end)
        """
        from datetime import timedelta
        
        current_date = start_date
        
        while current_date < end_date:
            train_start = current_date
            train_end = train_start + timedelta(days=self.training_window_days)
            test_start = train_end
            test_end = test_start + timedelta(days=self.test_window_days)
            
            if test_end > end_date:
                break
            
            yield train_start, train_end, test_start, test_end
            
            current_date += timedelta(days=self.step_days)


class RetrainingScheduler:
    """
    Schedules and manages model retraining.
    """
    
    def __init__(
        self,
        retrain_frequency_days: int = 30,
        min_trades_for_retrain: int = 50,
    ):
        """
        Initialize scheduler.
        
        Args:
            retrain_frequency_days: Retrain every N days
            min_trades_for_retrain: Minimum trades before retraining
        """
        self.retrain_frequency_days = retrain_frequency_days
        self.min_trades_for_retrain = min_trades_for_retrain
        self.last_retrain_date: Optional[datetime] = None
        self.trades_since_retrain: int = 0
    
    def should_retrain(self, current_date: datetime, num_new_trades: int = 1) -> bool:
        """
        Check if retraining should occur.
        
        Args:
            current_date: Current date
            num_new_trades: Number of new trades since last check
        
        Returns:
            True if retraining should occur
        """
        from datetime import timedelta
        
        self.trades_since_retrain += num_new_trades
        
        if self.last_retrain_date is None:
            return False
        
        days_since_retrain = (current_date - self.last_retrain_date).days
        
        # Retrain if either condition is met
        if days_since_retrain >= self.retrain_frequency_days:
            return True
        
        if self.trades_since_retrain >= self.min_trades_for_retrain:
            return True
        
        return False
    
    def mark_retrain_complete(self, current_date: datetime):
        """Mark that retraining has been completed."""
        self.last_retrain_date = current_date
        self.trades_since_retrain = 0


class SimpleAlphaTrainer(ModelTrainer):
    """
    Simple trainer for alpha models.
    Placeholder for actual ML training logic.
    """
    
    def __init__(self, model_type: str = "sklearn"):
        """
        Initialize trainer.
        
        Args:
            model_type: Type of model ("sklearn", "xgboost", "neural_net")
        """
        self.model_type = model_type
        self.model = None
    
    def train(self, training_data: Dict) -> object:
        """
        Train a simple classifier.
        
        Args:
            training_data: Dict with 'X' (features) and 'y' (labels)
        
        Returns:
            Trained model
        """
        if self.model_type == "sklearn":
            try:
                from sklearn.ensemble import RandomForestClassifier
                X = training_data.get('X', [])
                y = training_data.get('y', [])
                
                self.model = RandomForestClassifier(n_estimators=100, random_state=42)
                self.model.fit(X, y)
                
                return self.model
            except ImportError:
                return None
        
        return None
    
    def save_model(self, model: object, path: str):
        """Save model to disk."""
        with open(path, 'wb') as f:
            pickle.dump(model, f)
    
    def load_model(self, path: str) -> object:
        """Load model from disk."""
        with open(path, 'rb') as f:
            return pickle.load(f)


class PerformanceAnalyzer:
    """
    Analyzes trading performance to identify improvements.
    """
    
    @staticmethod
    def analyze_alpha_contribution(
        trades: List[Trade],
    ) -> Dict[str, Dict]:
        """
        Analyze which alphas contribute most to returns.
        
        Args:
            trades: List of trades
        
        Returns:
            Dict with per-alpha statistics
        """
        alpha_stats = {}
        
        for trade in trades:
            alpha = trade.alpha_name
            if alpha not in alpha_stats:
                alpha_stats[alpha] = {
                    'total_trades': 0,
                    'winning_trades': 0,
                    'total_pnl': 0.0,
                    'total_pnl_pct': 0.0,
                }
            
            stats = alpha_stats[alpha]
            stats['total_trades'] += 1
            if trade.pnl > 0:
                stats['winning_trades'] += 1
            stats['total_pnl'] += trade.pnl
            stats['total_pnl_pct'] += trade.pnl_pct
        
        # Calculate metrics
        for alpha, stats in alpha_stats.items():
            if stats['total_trades'] > 0:
                stats['win_rate'] = stats['winning_trades'] / stats['total_trades']
                stats['avg_return_pct'] = stats['total_pnl_pct'] / stats['total_trades']
        
        return alpha_stats
    
    @staticmethod
    def identify_regime_specific_alphas(
        trades: List[Trade],
    ) -> Dict:
        """
        Identify which alphas perform best in each regime.
        
        Returns:
            Dict with regime -> alpha performance mapping
        """
        # Placeholder - would need regime labels on trades
        return {}
