trading_agent/requirements.txt
pandas>=1.3.0
numpy>=1.20.0
pytz>=2021.1
pyyaml>=5.4.0
scikit-learn>=1.0.0

# Optional dependencies for extended features
xgboost>=1.5.0
matplotlib>=3.4.0
jupyter>=1.0.0

# For live broker integrations (optional)
# alpaca-trade-api>=0.60.0
# nsepy>=0.1
# ccxt>=1.60.0
